requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.8
django-cache-url==3.2.3
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-jalali-date==1.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.17.1
environs==8.0.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
jdatetime==3.6.4
Jinja2==3.0.3
kombu==5.2.2
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.14.1
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2==2.9.2
psycopg2-binary==2.8.5
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
skyroom==1.0.3
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
validate-email==1.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zipp==3.6.0
